--- frontend/src/services/api.js
@@ -1,111 +1,0 @@
-import axios from 'axios';
-
-const API_BASE_URL = '/api';
-
-// Create axios instance with default config
-const api = axios.create({
-  baseURL: API_BASE_URL,
-  timeout: 10000,
-});
-
-// Request interceptor to add auth token
-api.interceptors.request.use(
-  (config) => {
-    const token = localStorage.getItem('token');
-    if (token) {
-      config.headers.Authorization = `Bearer ${token}`;
-    } else {
-      console.error('No token found in localStorage');
-      // Redirect to login if no token
-      window.location.href = '/login';
-    }
-    return config;
-  },
-  (error) => {
-    return Promise.reject(error);
-  }
-);
-
-// Response interceptor to handle errors
-api.interceptors.response.use(
-  (response) => response,
-  (error) => {
-    if (error.response?.status === 401) {
-      // Token expired or invalid
-      localStorage.removeItem('token');
-      window.location.href = '/login';
-    }
-    return Promise.reject(error);
-  }
-);
-
-// Expenses API
-export const expensesAPI = {
-  getAll: (params) => api.get('/expenses', { params }),
-  getById: (id) => api.get(`/expenses/${id}`),
-  create: (data) => {
-    const formData = new FormData();
-    
-    // Ajouter tous les champs au FormData
-    Object.keys(data).forEach(key => {
-      if (data[key] !== null && data[key] !== undefined && data[key] !== '') {
-        formData.append(key, data[key]);
-      }
-    });
-    
-    return api.post('/expenses', formData, {
-      headers: { 
-        'Content-Type': 'multipart/form-data'
-      }
-    });
-  },
-  update: (id, data) => {
-    const formData = new FormData();
-    Object.keys(data).forEach(key => {
-      if (data[key] !== null && data[key] !== undefined && data[key] !== '') {
-        formData.append(key, data[key]);
-      }
-    });
-    return api.put(`/expenses/${id}`, formData, {
-      headers: { 'Content-Type': 'multipart/form-data' }
-    });
-  },
-  delete: (id) => api.delete(`/expenses/${id}`)
-};
-
-// Incomes API
-export const incomesAPI = {
-  getAll: (params) => api.get('/incomes', { params }),
-  getById: (id) => api.get(`/incomes/${id}`),
-  create: (data) => api.post('/incomes', data),
-  update: (id, data) => api.put(`/incomes/${id}`, data),
-  delete: (id) => api.delete(`/incomes/${id}`)
-};
-
-// Categories API
-export const categoriesAPI = {
-  getAll: () => api.get('/categories'),
-  create: (data) => api.post('/categories', data),
-  update: (id, data) => api.put(`/categories/${id}`, data),
-  delete: (id) => api.delete(`/categories/${id}`)
-};
-
-//Summary API
-export const summaryAPI = {
-  getMonthly: (params) => api.get('/summary/monthly', { params }),
-  getCustom: (params) => api.get('/summary', { params }),
-  getAlerts: () => api.get('/summary/alerts')
-};
-
-
-// User API
-export const userAPI = {
-  getProfile: () => api.get('/user/profile')
-};
-
-// Receipts API
-export const receiptsAPI = {
-  getReceipt: (id) => api.get(`/receipts/${id}`, { responseType: 'blob' })
-};
-
-export default api;
