+++ frontend/src/services/api/config.js
@@ -1,0 +1,4 @@
+export const API_CONFIG = {
+  BASE_URL: '/api',
+  TIMEOUT: 10000
+};
