+++ frontend/src/services/categoriesAPI.js
@@ -1,0 +1,8 @@
+import api from './api';
+
+export const categoriesAPI = {
+  getAll: () => api.get('/categories'),
+  create: (data) => api.post('/categories', data),
+  update: (id, data) => api.put(`/categories/${id}`, data),
+  delete: (id) => api.delete(`/categories/${id}`)
+};
