+++ frontend/src/services/incomesAPI.js
@@ -1,0 +1,9 @@
+import api from './api';
+
+export const incomesAPI = {
+  getAll: (params) => api.get('/incomes', { params }),
+  getById: (id) => api.get(`/incomes/${id}`),
+  create: (data) => api.post('/incomes', data),
+  update: (id, data) => api.put(`/incomes/${id}`, data),
+  delete: (id) => api.delete(`/incomes/${id}`)
+};
