+++ frontend/src/services/index.js
@@ -1,0 +1,7 @@
+export { expensesAPI } from './expensesAPI';
+export { incomesAPI } from './incomesAPI';
+export { categoriesAPI } from './categoriesAPI';
+export { summaryAPI } from './summaryAPI';
+export { userAPI } from './userAPI';
+export { receiptsAPI } from './receiptsAPI';
+export { default as api } from './api';
