+++ frontend/src/services/receiptsAPI.js
@@ -1,0 +1,5 @@
+import api from './api';
+
+export const receiptsAPI = {
+  getReceipt: (id) => api.get(`/receipts/${id}`, { responseType: 'blob' })
+};
