+++ frontend/src/services/summaryAPI.js
@@ -1,0 +1,7 @@
+import api from './api';
+
+export const summaryAPI = {
+  getMonthly: (params) => api.get('/summary/monthly', { params }),
+  getCustom: (params) => api.get('/summary', { params }),
+  getAlerts: () => api.get('/summary/alerts')
+};
