+++ frontend/src/services/userAPI.js
@@ -1,0 +1,5 @@
+import api from './api';
+
+export const userAPI = {
+  getProfile: () => api.get('/user/profile')
+};
